// src/api/models/process.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Process
 */
export interface Process {
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'version'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'status'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'deploymentTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'processDefinitionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'startTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'currentTask'?: string;
    /**
     * 
     * @type {string}
     * @memberof Process
     */
    'assignee'?: string;
}